Renamed and removed a few files
