+++ templates/pages/statistics.ejs
@@ -60,8 +60,6 @@
 
 <script src="./js/download.js"></script>
 <script src="./js/Chart.bundle.min.js"></script>
-<script src="./js/statistics2.js" type="module"></script>
-<!-- <script src="./js/d3.min.js"></script>
-<script src="./js/datavis.js"></script> -->
+<script src="./js/statistics.js" type="module"></script>
 
 <% include ../partials/tail %>
